Rename root packages
